remove code duplication
