+++ ra-optimisation/search-tree-rao/src/main/java/com/powsybl/openrao/searchtreerao/marmot/results/extensions/JsonPreTimeCouplingOverloadedCnecs.java
@@ -7,67 +7,28 @@
 
 package com.powsybl.openrao.searchtreerao.marmot.results.extensions;
 
-import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonToken;
 import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.SerializerProvider;
 import com.google.auto.service.AutoService;
-import com.powsybl.openrao.data.raoresult.io.json.RaoResultJsonConstants;
 import com.powsybl.openrao.data.raoresult.io.json.RaoResultJsonUtils;
+import com.powsybl.openrao.data.raoresult.io.json.extension.AbstractJsonCnecIdsExtension;
 
 import java.io.IOException;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
 
 /**
  * @author Peter Mitri {@literal <peter.mitri at rte-france.com>}
  */
 @AutoService(RaoResultJsonUtils.ExtensionSerializer.class)
-public class JsonPreTimeCouplingOverloadedCnecs implements RaoResultJsonUtils.ExtensionSerializer<PreTimeCouplingOverloadedCnecs> {
-
-    @Override
-    public void serialize(PreTimeCouplingOverloadedCnecs criticalCnecsResult, JsonGenerator jsonGenerator, SerializerProvider serializerProvider) throws IOException {
-        List<String> sortedListofFlowCnecIds = criticalCnecsResult.getCriticalCnecIds().stream().sorted().toList();
-
-        jsonGenerator.writeStartArray();
-        for (String consideredCnec : sortedListofFlowCnecIds) {
-            jsonGenerator.writeStartObject();
-            jsonGenerator.writeStringField(RaoResultJsonConstants.FLOWCNEC_ID, consideredCnec);
-            jsonGenerator.writeEndObject();
-        }
-        jsonGenerator.writeEndArray();
-
-    }
+public class JsonPreTimeCouplingOverloadedCnecs extends AbstractJsonCnecIdsExtension<PreTimeCouplingOverloadedCnecs> {
 
     @Override
     public PreTimeCouplingOverloadedCnecs deserialize(JsonParser jsonParser, DeserializationContext deserializationContext) throws IOException {
         return deserializeAndUpdate(jsonParser, deserializationContext, new PreTimeCouplingOverloadedCnecs());
     }
 
-    public PreTimeCouplingOverloadedCnecs deserializeAndUpdate(JsonParser jsonParser, DeserializationContext deserializationContext, PreTimeCouplingOverloadedCnecs criticalCnecsResult) throws IOException {
-
-        Set<String> criticalCnecsIdSet = new HashSet<>();
-        while (jsonParser.nextToken() != JsonToken.END_ARRAY) {
-            if (jsonParser.nextFieldName().equals(RaoResultJsonConstants.FLOWCNEC_ID)) {
-                criticalCnecsIdSet.add(jsonParser.nextTextValue());
-            }
-            jsonParser.nextToken();
-        }
-        criticalCnecsResult.setCriticalCnecIds(criticalCnecsIdSet);
-
-        return criticalCnecsResult;
-    }
-
     @Override
     public String getExtensionName() {
         return "pre-time-coupling-overloaded-cnecs";
-    }
-
-    @Override
-    public String getCategoryName() {
-        return "rao-result";
     }
 
     @Override
